Updating Code

+++ app/Http/Controllers/UserController.php
@@ -10,6 +10,7 @@ use Gate;
 use Illuminate\Support\Facades\Cache;
 use Illuminate\Support\Facades\Log;
 use App\UserSettings;
+use Illuminate\Support\Facades\Auth;
 
 class UserController extends Controller
 {
@@ -17,6 +18,7 @@ class UserController extends Controller
     {
         if (Gate::denies('view-acp')) {
             toast(__('site.permission_denied'), 'warning');
+            Log::channel('app')->notice("[PERMISSION DENIED] User " . auth()->user()->name . " (ID: " . auth()->user()->id . ") attempted to access " . request()->path());
             return redirect()->route('home');
         }
 
@@ -34,6 +36,11 @@ class UserController extends Controller
 
         if (!$user) {
             toast(__('user.invalid_user'), 'error');
+            if (auth()->user()) {
+                Log::channel('app')->warning("[Invalid User] User " . auth()->user()->name . " (ID: " . auth()->user()->id . ") attempted to access " . request()->path() . " but the user does not exist.");
+            } else {
+                Log::channel('app')->warning("[Invalid User] User GUEST (IP: " . request()->ip() . ") attempted to access " . request()->path() . " but the user does not exist.");
+            }
             return redirect()->route('home');
         }
 
@@ -48,10 +55,17 @@ class UserController extends Controller
 
     public function edit($id)
     {
+        if (Gate::denies('edit-user', $id)) {
+            toast(__('site.permission_denied'), 'warning');
+            Log::channel('app')->notice("[PERMISSION DENIED] User " . auth()->user()->name . " (ID: " . auth()->user()->id . ") attempted to access " . request()->path());
+            return redirect()->route('home');
+        }
+
         $user = $this->getUser($id);
 
         if (!$user) {
             toast(__('user.invalid_user'), 'error');
+            Log::channel('app')->warning("[Invalid User] User " . auth()->user()->name . " (ID: " . auth()->user()->id . ") attempted to access " . request()->path() . " but the user does not exist.");
             return redirect()->route('home');
         }
 
@@ -78,10 +92,17 @@ class UserController extends Controller
 
     public function update(Request $request, $id)
     {
+        if (Gate::denies('edit-user', $id)) {
+            toast(__('site.permission_denied'), 'warning');
+            Log::channel('app')->notice("[PERMISSION DENIED] User " . auth()->user()->name . " (ID: " . auth()->user()->id . ") attempted to access " . request()->path());
+            return redirect()->route('home');
+        }
+
         $user = $this->getUser($id);
 
         if (!$user) {
             toast(__('user.invalid_user'), 'error');
+            Log::channel('app')->warning("[Invalid User] User " . auth()->user()->name . " (ID: " . auth()->user()->id . ") attempted to access " . request()->path() . " but the user does not exist.");
             return redirect()->route('home');
         }
 
@@ -90,7 +111,7 @@ class UserController extends Controller
             'pushover_key' => 'nullable|string',
         ]);
 
-        Log::channel('app')->info("[User Update] id: " . $user->id . " original: " . json_encode($user) . " new: " . json_encode($request->all()));
+        Log::channel('app')->info("[User Update] id: " . $user->id . " updated by " . auth()->user()->name . " (ID: " . auth()->user()->id . ") original: " . json_encode($user) . " Original User Settings: " . json_encode($user->settings) . " new: " . json_encode($request->all()));
 
         $topics = $this->getContactTopics();
 
@@ -154,6 +175,7 @@ class UserController extends Controller
     {
         if (Gate::denies('manage-user-roles')) {
             toast(__('site.permission_denied'), 'warning');
+            Log::channel('app')->notice("[PERMISSION DENIED] User " . auth()->user()->name . " (ID: " . auth()->user()->id . ") attempted to access " . request()->path());
             return redirect()->route('home');
         }
 
@@ -161,6 +183,7 @@ class UserController extends Controller
 
         if (!$user) {
             toast(__('user.invalid_user'), 'error');
+            Log::channel('app')->warning("[Invalid User] User " . auth()->user()->name . " (ID: " . auth()->user()->id . ") attempted to access " . request()->path() . " but the user does not exist.");
             return redirect()->route('home');
         }
 
@@ -172,6 +195,7 @@ class UserController extends Controller
 
         if (!$role) {
             toast(__('user.invalid_role'), 'error');
+            Log::channel('app')->warning("[Invalid Role] User " . auth()->user()->name . " (ID: " . auth()->user()->id . ") attempted to access " . request()->path() . " with role " . $request->role . " but the role does not exist.");
             return redirect()->route('profile', $user->id);
         }
 
@@ -183,7 +207,7 @@ class UserController extends Controller
 
         $user->roles()->attach($request->role);
 
-        Cache::forget('user:' . $user->id . ':is:' . $role->name);
+        Cache::forget('User:' . $user->id . ':is:' . $role->name);
 
         toast(__('user.role_add_success'), 'success');
 
@@ -194,6 +218,7 @@ class UserController extends Controller
     {
         if (Gate::denies('manage-user-roles')) {
             toast(__('site.permission_denied'), 'warning');
+            Log::channel('app')->notice("[PERMISSION DENIED] User " . auth()->user()->name . " (ID: " . auth()->user()->id . ") attempted to access " . request()->path());
             return redirect()->route('home');
         }
 
@@ -205,8 +230,6 @@ class UserController extends Controller
         $confirm['action'] = route("remove-role-confirm", ['id' => $user->id, 'role' => $role->id]);
         $confirm['cancel'] = route('profile', $user->id);
 
-        Cache::forget('user:' . $user->id . ':is:' . $role->name);
-
         return view('site.confirm', [
             'user' => $user,
             'confirm_data' => $confirm,
@@ -215,12 +238,22 @@ class UserController extends Controller
 
     public function delRoleConfirm(Request $request, $id, $role_id)
     {
+        if (Gate::denies('manage-user-roles')) {
+            toast(__('site.permission_denied'), 'warning');
+            Log::channel('app')->notice("[PERMISSION DENIED] User " . auth()->user()->name . " (ID: " . auth()->user()->id . ") attempted to access " . request()->path());
+            return redirect()->route('home');
+        }
+
         $user = $this->getUser($id);
         $role = $this->getRole($role_id);
 
         $user->roles()->detach($role->id);
 
+        Cache::forget('User:' . $user->id . ':is:' . $role->name);
+
         toast(__('user.role_del_success'), 'success');
+        
+        Log::channel('app')->info("[Role Remove] User " . auth()->user()->name . " (ID: " . auth()->user()->id . ") removed role " . $role->name . " from User " . $user->name . " (ID: " . $user->id . ")");
 
         return redirect()->route('profile', $user->id);
     }
